requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
numpy
opencv-python-headless
google-generativeai
python-dotenv